Move calculateTopLevelDecision to rules
