+++ src/decision.ts
@@ -3,25 +3,6 @@ import { CardefPool } from './pool';
 import { StateManager } from './state';
 import { CardType, Zone } from './types';
 
-export function calculateTopLevelDecision(state: State): Decision {
-    const stateManager = new StateManager(state);
-    const mySideManager = stateManager.getMySideManager();
-    const hand = mySideManager.hand;
-    // ToDo: Don't offer hand cards that can't be played or discarded
-    const availableSlots = hand.map((card, index) => {
-        const slot: Slot = {
-            zone: Zone.MY_HAND,
-            index,
-        };
-        return slot;
-    });
-    return {
-        label: 'Play/discard/harvest/fight with which card?',
-        availableSlots,
-        selectedSlots: [],
-    };
-}
-
 export function getTopLevelSlot(state: State): Slot {
     const topLevelDecision = state.currentDeed.decisions[0];
     if (!topLevelDecision) {
